+++ app/src/main/java/ru/netology/nmedia/dao/PostRemoteKeyDao.kt
@@ -10,10 +10,10 @@ import ru.netology.nmedia.entity.PostRemoteKeyEntity
 interface PostRemoteKeyDao {
 
     @Query("SELECT max(`key`) FROM PostRemoteKeyEntity")
-    suspend fun max(): Long?
+    suspend fun max(): Int?
 
     @Query("SELECT min(`key`) FROM PostRemoteKeyEntity")
-    suspend fun min(): Long?
+    suspend fun min(): Int?
 
     @Insert(onConflict = OnConflictStrategy.REPLACE)
     suspend fun insert(postRemoteKeyEntity: PostRemoteKeyEntity)
